requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM dependencies
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# pgvector support
pgvector==0.2.4

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.0.3

# Image processing
Pillow==10.1.0

# Environment and configuration
python-dotenv==1.0.0

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# Utilities
python-dateutil==2.8.2